Add support parameterized LIMIT in the sub plan with volatile functions
If the LimitPath has a General locus and the subplan contains a volatile
function, then Motion can be added on top of the Limit at a later stage
of processing. In this case, if Limit is parameterized, it produces
incorrect output because the limit parameters do not pass through Motion.
To prevent this, we calculate the data set to be processed in the limit on
one segment and send it to the remaining ones. To do this, we set the
underlying Limit node to locus SingleQE, and set the Limit node to OuterQuery
(since it contains parameters) and add Motion between them.
If LimitPath has an OuterQuery, we also add Motion so that the data set for
Limit processing on all segments is identical.